+++ go.mod
@@ -23,6 +23,7 @@ require (
 	github.com/stretchr/objx v0.1.2-0.20180626195558-9e1dfc121bca // indirect
 	github.com/stretchr/testify v1.6.1
 	golang.org/x/crypto v0.0.0-20200220183623-bac4c82f6975 // indirect
+	gopkg.in/yaml.v2 v2.2.8
 	k8s.io/api v0.16.10
 	k8s.io/apimachinery v0.16.10
 	k8s.io/client-go v0.15.12
